added new gmail admin

+++ app/main/routes.py
@@ -13,12 +13,13 @@ from redbeat import RedBeatSchedulerEntry
 
 from . import main
 from .. import db, celery
+from app.email import send_email
 from ..models import Reminder
 
 class ReminderForm(FlaskForm):
     subject = StringField(label='Subject', description='What should be the mail subject?', validators=[Optional(), Length(1,50, 'Subject must have between 5 and 50 characters.')] )
     content = TextAreaField(label='Content', description='What should be the mail content?', validators=[Optional(), Length(1,700, 'Content must have between 1 and 200 characters.')])
-    days = SelectMultipleField(label='Days', description='What days of the week should I email you?', validators=[Optional()], choices = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'], default = None)
+    days = SelectMultipleField(label='Days', description='What days of the week should I email you?', validators=[Optional()], choices = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'], render_kw = {'class': 'days-select'},  default = None)
     # TODO: only allow dates past today
     date = DateField(label='Date', description='When should I email you?', validators=[Optional()], default = None)
     time = TimeField(label='Time', description='What time should I email you?', validators=[DataRequired('A time is required.')], default=datetime.time(0,0))
@@ -27,8 +28,11 @@ class ReminderForm(FlaskForm):
 @main.route('/')
 def index():
     if current_user.is_authenticated:
-        reminders = Reminder.query.filter_by(author_id = current_user.id).all()
-        return render_template('home.html', reminders = reminders)
+        if current_user.confirmed:
+            reminders = Reminder.query.filter_by(author_id = current_user.id).all()
+            return render_template('home.html', reminders = reminders)
+        else: 
+            return render_template('confirm.html')
     else:
         return redirect('/auth/register')
 
@@ -52,27 +56,32 @@ def new(reminder_type):
         db.session.add(r)
         db.session.commit()
 
+        user_tz = pytz.timezone(current_user.timezone)
+        
         if reminder_type == 'periodic':
+            print(current_user.timezone)
             interval = schedules.crontab(
                 minute = data['time'].minute,
                 hour = data['time'].hour,
                 day_of_week = list_to_string(data['days']),
-                nowfun = partial(datetime.datetime.now, pytz.timezone('America/Argentina/Buenos_Aires'))
-                # TODO: add user timezone as column. ask for it in register page
+                nowfun = partial(datetime.datetime.now, user_tz)
             )
             entry = RedBeatSchedulerEntry(
                 name = str(r.id),
                 task = 'app.email.send_email',
                 schedule = interval,
-                args = (r.id, current_user.email, r.subject, r.content),
+                args = (current_user.email, r.subject, r.content),
+                kwargs = ({'reminder_id': r.id}),
                 app = celery
             )
-            print(entry.key)
             entry.save()
-        
-        print(r.id)
-        # TODO: save entry key as database column
-        
+        elif reminder_type == 'deadline':
+            send_email.apply_async(
+                task_id = str(r.id),
+                args = (current_user.email, r.subject, r.content),
+                kwargs = ({'reminder_id': r.id}),
+                eta = user_tz.localize(datetime.datetime.combine(r.date, r.time))
+            )
         # TODO: implement deadline reminders
         return redirect('/')
         
@@ -88,7 +97,6 @@ def edit(reminder_id):
             days = r.days,
             time = r.time
             )
-    form.submit = SubmitField('Edit')
     if form.validate_on_submit():
         data = form.data
         stmt = (
@@ -110,11 +118,16 @@ def edit(reminder_id):
 @main.route('/delete/<reminder_id>', methods=['POST'])
 def delete(reminder_id):
     r = Reminder.query.filter_by(id = reminder_id).first()
+    
+    if r.days != None: # periodic reminder
+        e = RedBeatSchedulerEntry.from_key(f'redbeat:{r.id}', app=celery)
+        e.delete()
+    elif r.date != None: # deadline reminder
+        celery.control.revoke(str(r.id))
+    
     db.session.delete(r)
     db.session.commit()
     
-    e = RedBeatSchedulerEntry.from_key(f'redbeat{r.id}')
-    e.delete()
     return redirect('/')
